Fixed all the bugs i could find
I am working on a new function that prevents you from making a move that
is impossible (if you want to start a ship somewhere but there would be
no possible outcomes). it prevents right now but does not pop up my
alert box saying the move is impossible. will continue to work on this.